requirements.txt
google-generativeai
altair==4.0
streamlit==1.29.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.1
googleapis-common-protos==1.62.0
python-dotenv